rename schema type objects for clarity
